// src/db/schema.ts
// Auto-generated by Crudora — do not edit manually
import { datetime, mysqlTable, varchar } from 'drizzle-orm/mysql-core';

export const userTable = mysqlTable('user', {
  id: varchar('id', { length: 36 }).primaryKey(),
  name: varchar('name', { length: 255 }).notNull(),
  email: varchar('email', { length: 255 }).notNull().unique(),
  password: varchar('password', { length: 255 }).notNull(),
  createdAt: datetime('createdAt', { mode: 'date' }).notNull(),
  updatedAt: datetime('updatedAt', { mode: 'date' }).notNull(),
  deletedAt: datetime('deletedAt', { mode: 'date' }),
});